requirements.txt
selenium==4.8.0
beautifulsoup4==4.11.1
jsonpickle==3.0.1
openai==0.26.4
discord.py==2.1.0
black==22.3.0
flake8==5.0.4
isort==5.10.1

